+++ trs/testsettings.py
@@ -66,7 +66,7 @@ Logging = {
         },
          'loggers': {
              '': {
-                 'handlers': [],
+                 'handlers': ['console', 'logfile'],
                  'propagate': True,
                  'level': 'DEBUG',
              },
@@ -87,3 +87,6 @@ TRS_END_YEAR = 2020
 USE_L10N = True
 USE_I18N = True
 LANGUAGE_CODE = 'nl-nl'
+
+LOGIN_URL = 'trs.login'
+LOGOUT_URL = 'trs.logout'