+++ trs/views.py
@@ -1,8 +1,14 @@
 import datetime
 import logging
 
+from django.utils.decorators import method_decorator
 from django import forms
+from django.contrib.auth import logout, login, authenticate
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.forms import AuthenticationForm
+from django.core.urlresolvers import reverse
 from django.db import models
+from django.shortcuts import redirect
 from django.utils.datastructures import SortedDict
 from django.views.generic.base import TemplateView
 from django.views.generic.edit import FormView
@@ -15,6 +21,14 @@ from trs.models import WorkAssignment
 
 
 logger = logging.getLogger()
+
+
+class LoginRequiredMixin(object):
+    """See http://stackoverflow.com/a/10304880/27401"""
+
+    @method_decorator(login_required)
+    def dispatch(self, *args, **kwargs):
+        return super(LoginRequiredMixin, self).dispatch(*args, **kwargs)
 
 
 class BaseMixin(object):
@@ -44,8 +58,9 @@ class BaseMixin(object):
         return self.active_person.assigned_projects()
 
 
-class BaseView(TemplateView, BaseMixin):
+class BaseView(LoginRequiredMixin, TemplateView, BaseMixin):
     pass
+
 
 class HomeView(BaseView):
     template_name = 'trs/home.html'
@@ -83,7 +98,28 @@ class ProjectView(BaseView):
         return Project.objects.get(slug=self.kwargs['slug'])
 
 
-class BookingView(FormView, BaseMixin):
+class LoginView(FormView, BaseMixin):
+    template_name = 'trs/login.html'
+    form_class = AuthenticationForm
+
+    @property
+    def success_url(self):
+        return reverse('trs.booking')
+
+    def form_valid(self, form):
+        username = form.cleaned_data['username']
+        password = form.cleaned_data['password']
+        user = authenticate(username=username, password=password)
+        login(self.request, user)
+        return super(LoginView, self).form_valid(form)
+
+
+def logout_view(request):
+    logout(request)
+    return redirect('trs.home')
+
+
+class BookingView(LoginRequiredMixin, FormView, BaseMixin):
     # TODO: also allow /booking/yyyy-ww/ format.
     template_name = 'trs/booking.html'